pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.28.0",
    "pdfplumber>=0.11.7",
    "reportlab>=4.4.3",
    "streamlit>=1.47.1",
]
